refactor(api): rename files
